requirement.txt
gTTS
playsound
pytesseract
Pillow
pykakasi
pynput
pyperclip